requirements.txt
beautifulsoup4==4.5.3
boto==2.40.0
colour==0.1.2
contextlib2==0.5.3
dj-database-url==0.4.1
Django==1.10.5
django-modelcluster==2.0
django-storages==1.4.1
django-taggit==0.18.3
django-treebeard==4.1.0
djangorestframework==3.5.3
graphene==1.2
graphene-django==1.2.1
graphql-core==1.0.1
graphql-relay==0.4.5
gunicorn==19.6.0
html5lib==0.999999999
iso8601==0.1.11
olefile==0.44
palette==0.2
Pillow==4.0.0
promise==1.0.1
psycopg2==2.6.1
pytz==2016.10
raven==5.19.0
requests==2.12.5
singledispatch==3.4.0.3
six==1.10.0
typing==3.6.1
Unidecode==0.4.20
wagtail==1.8
webencodings==0.5
whitenoise==3.2
Willow==0.4
